[Refactor] #22 - redis pub/sub dir 이동
